+++ test/PKAST.m
@@ -77,7 +77,7 @@
 }
 
 
-- (NSInteger)type {
+- (int)type {
     NSAssert2(0, @"%s is an abastract method. Must be overridden in %@", __PRETTY_FUNCTION__, NSStringFromClass([self class]));
     return -1;
 }
